Sweeping changes including use of react-dropzone
