+++ dev/actions/index.js
@@ -1,3 +1,2 @@
 export * from './authentication'
 export * from './cattle'
-export * from './images'
