+++ dev/components/Cattle/CattleIndexComponent.jsx
@@ -1,4 +1,4 @@
-require('./CattleComponent.scss');
+require('./CattleIndexComponent.scss');
 
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
