--- dev/components/Cattle/CattleRecognitionComponent.scss
@@ -1,26 +1,0 @@
-@import "../../style/colours.scss";
-@import "../../style/fonts.scss";
-
-.cattle-recognition-component-wrapper {
-  margin: auto;
-  max-width: 600px;
-  text-align: center;
-}
-
-.cattle-recognition-component-title {
-  font-size: $font-size-bigger;
-}
-
-.cattle-recognition-component-input {
-  box-sizing: border-box;
-  border: 1px solid $background-colour-dark;
-  border-radius: 2px;
-  font-family: $font-family-default;
-  margin: 10px 20px;
-  padding: 8px 6px;
-  width: 100%;
-}
-
-.cattle-recognition-component-button-submit {
-  padding: 6px 10px;
-}
