--- dev/components/Cattle/DropzoneComponent.scss
@@ -1,36 +1,0 @@
-@import "../../style/colours.scss";
-@import "../../style/fonts.scss";
-
-.dropzone-component-wrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-%dropzone-component-dropzone {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 60vw;
-  width: 60vw;
-}
-
-.dropzone-component-dropzone {
-  @extend %dropzone-component-dropzone;
-  border: 2px dashed #666;
-}
-
-.dropzone-component-wrapper-dropzone {
-  @extend %dropzone-component-dropzone;
-  border: 2px solid #666;
-}
-
-.dropzone-component-notice {
-  font-size: $font-size-big;
-  color: #666;
-}
-
-.dropzone-component-image-preview {
-  height: 60vw;
-  width: 60vw;
-}
