+++ dev/style/layout.scss
@@ -22,6 +22,7 @@ body {
 }
 
 button {
+  background-color: $background-colour;
   border: 0;
   cursor: pointer;
   margin: auto;
