+++ package.json
@@ -35,6 +35,7 @@
     "react": "^15.3.2",
     "react-dom": "^15.3.2",
     "react-dropdown": "^1.1.0",
+    "react-dropzone": "^3.7.3",
     "react-redux": "^4.4.5",
     "react-router": "^2.8.1",
     "react-router-redux": "^4.0.6",
